+++ Mods ChoGGi/Nursery Upgrade Slots/metadata.lua
@@ -1,0 +1,21 @@
+return PlaceObj("ModDef", {
+	"title", "Nursery: Upgrade Slots",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 1557057600,
+	"image", "Preview.png",
+	"id", "ChoGGi_NurseryUpgradeSlots",
+	"steam_id", "1733456204",
+	"pops_any_uuid", "74b2cdfe-5c31-4fbf-a563-4afa1f1580d0",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision or 244275,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Adds upgrades to change the slot capacity to 16 32 64 (also increases electricity consumption).
+The second/third upgrades will be unlocked after the first/second is built once.
+
+
+Requested by Nacira.]],
+})