+++ Expanded Cheat Menu/Changelog.md
@@ -27,8 +27,11 @@
 - Mod Upload supports batch uploading (and screenshots).
 - Some menu items now use translated text instead of mine.
 - Debug>Examine will try to examine the UI element under the cursor if it can't find an object.
-- Console>~obj will show all results instead of only the first one (in a table).
 - Renamed imgview/txtview to OpenImageViewer/OpenTextViewer.
+- Cleaned up Debug>View All Entities.
+- Console:
+- > ~obj will show all results instead of only the first one (in a table).
+- > Right-clicking history or scripts will paste the code into the console.
 - Menu items to use the default ECM icon instead of something maybe vaguely related:
 - > Easier to distinguish between notifications from ECM and from SM.
 - Examine: